fix: restored missing tests
